695 - refactor
